plugins/fpr-editorial-agent/requirements.txt
lxml>=4.9.0
pyyaml>=6.0
click>=8.1.0
defusedxml>=0.7.0
